+++ api/types/stats.go
@@ -1,6 +1,6 @@
 // This package is used for API stability in the types and response to the
 // consumers of the API stats endpoint.
-package stats
+package types
 
 import "time"
